[ADD] the auth middleware for the post, fix some controllers in post

+++ groupomania/back/controllers/post.js
@@ -1,5 +1,6 @@
 const db = require('../models');
 const fs = require('fs');
+const auth = require('../middleware/auth');
 
 const Post = db.post;
 
@@ -7,10 +8,13 @@ exports.create = (req, res, next) => {
   const postObjectTemp = JSON.stringify(req.body); 
   const postObject = JSON.parse(postObjectTemp);  //très très très mauvaise pratique
   delete postObject._id;
-  const post = new Post({
+  let post = new Post({
     ...postObject,
-    image: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    author: req.userId,
   });
+  if (req.file) {
+    post.image = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`;
+  }
   post.save()
     .then(() => res.status(201).json({ message: 'Post crée !' }))
     .catch(error => res.status(400).json({ error }));
@@ -31,12 +35,14 @@ exports.getOne = (req, res, next) => {
 exports.deleteOne = (req, res, next) => {
   Post.findOne({where:{ id: req.params.id }})
     .then(Post => {
-      const imageName = Post.image.split('/images/')[1];
-      fs.unlink(`images/${imageName}`, (error) => {
-        if (error) {
-          return res.status(500).json({ error: 'Erreur lors de la suppression de l\'image' });
-        }
-      })
+      if (Post.image !== null) {
+        const imageName = Post.image.split('/images/')[1];
+        fs.unlink(`images/${imageName}`, (error) => {
+          if (error) {
+            return res.status(500).json({ error: 'Erreur lors de la suppression de l\'image' });
+          }
+        })
+      }
       Post.destroy({where:{ id: req.params.id }})
         .then(() => res.status(200).json({ message: 'Post supprimé !' }))
         .catch(error => res.status(400).json({ error }));
